# Attempt faster datetime parsing with backports.datetime_fromisoformat