--- a/setup.cfg
+++ b/setup.cfg
@@ -31,6 +31,7 @@ package_dir =
     = src
 packages = find:
 install_requires =
+    backports-datetime-fromisoformat==2.0.0
     ccxt==2.5.1
     Historic-Crypto>=0.1.6
     jsonschema>=3.2.0